Fix to not depend on closure environment
The added closure keeps from introducing `global` as a global variable in browsers and shells. Which makes this more portable.

Also, testing that `this` is an object works better when loaded from a script context, as creating a function is unnecessary unless in a module context. Because of that, this is CSP safe unless the global scope is in strict mode already (which is bad practice from the start).